requirements.txt
redis
pika~=1.3.1
python-dotenv~=0.20.0